Add files via upload
This project was for CS-303 (Data Structures). The assignment was to write a banking algorithm that will collect customer information as it comes in, and then serve the last customer first. To do this, I had to utilize stacks to collect different types of data to use. I consider this project a culmination of everything I have learned in CS-101 and CS-201.